+++ scripts/afterSign.js
@@ -9,6 +9,11 @@ const { notarize } = require('electron-notarize')
 exports.default = async function (context) {
   if (context.electronPlatformName !== 'darwin') return
 
+  if (!process.env.APPLE_ID || !process.env.APPLE_ID_PASS) {
+    console.log('Skipping notarization: Either APPLE_ID or APPLE_ID_PASS is not in process.env.')
+    return
+  }
+
   // Retrieve the final application filename
   const appName = context.packager.appInfo.productFilename