+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 - Added an additional check to make sure to differ between explicit and implicit paste events in the CodeMirror instance.
 - Finally fixed the weird glitches of the file list. Now it will correctly scroll files into view, not break, and be not empty for a fraction of a second.
 - Overhauled the QuickLook windows. Now they react to much more finetuned configuration changes, are more responsive, and in general react faster to changes.
+- Switch to Electron 8.
 
 # 1.5.0